+++ Photo Club Hub/View/SinglePortFolioView/SinglePortfolioView.swift
@@ -28,8 +28,9 @@ struct SinglePortfolioView: UIViewRepresentable {
 
 // only visible when canvas is simulating/running
 struct SinglePortfolioView_Previews: PreviewProvider {
+
     static var webView = WKWebView()
-    static var url: URL = URL(string: FotogroepWaalreMembersProvider.baseURL + "Peter_van_den_Hamer/")!
+    static var url: URL = URL(string: "http://www.vdHamer.com/fgWaalre/Peter_van_den_Hamer/")!
 
     static var previews: some View {
         NavigationStack {
@@ -39,4 +40,5 @@ struct SinglePortfolioView_Previews: PreviewProvider {
                 .navigationBarTitleDisplayMode(.large)
         }
     }
+
 }